Update 03_pandas.ipynb
Line 92 is the code I used to find the age of the youngest respondent.
Line 119 is the code used to find the age of person with most number of pets.
Line 174 is for the third question of most popular pet.
For line 219, I do not have an answer, and I need to study more to figure out the lambda and pandas with multiple conditions. I do not understand the "series" function.